refactor(language): extract intelligence module
